fix [Workflow] A combined algorithm (training-prediction) is configured in the workflow, and the prediction algorithm fails to run

+++ rosettaflow-mapper/src/main/java/com/moirae/rosettaflow/mapper/WorkflowNodeMapper.java
@@ -14,5 +14,5 @@ import java.util.List;
  */
 public interface WorkflowNodeMapper extends BaseMapper<WorkflowNode> {
 
-    List<WorkflowNode> queryWorkflowNodeAndStatusList(@Param("workflowId") Long workflowId,  @Param("editVersion") Long editVersion);
+    List<WorkflowNode> queryWorkflowNodeAndStatusList(@Param("workflowId") Long workflowId,  @Param("editVersion") Integer editVersion);
 }
